Experiment to see if an alternate loader will work
There are 2/3 problems with the current webpack system:

1. Each entry point means a new webpack build
2. Each webpack build duplicates a huge number of resources
3. It's possible that we could run into problems with assumptions about
   what module level sharing happens

The first 2 points are emphasised when you consider that there are
currently 180+ script tags spread across 103 html/xhtml/xul files.

So what if we have a web-server that filters the resources that it
delivers to the browser with the intent that the current set of script
tags just works?

This would mean that modules would need to be bablified, AMD module
wrapped, and (if they are referenced by script tag rather than require),
also have requirejs and babel-polyfill shims prepended.

So this commit does the, above. And it mostly works. There are a couple
of drawbacks, neither fatal, which is why I've not persued it further so
far.

1. It's slow. Loading one large script file is faster than loading many.
   The solution to this problem is fairly simple: H2.

2. With the code as it stands, there is a problem with 'Arg' not being
   defined. The obvious explanation is a module loop, but it's not clear
   how that is happening.

This commit requires some changes to the node-ecstatic library. I've
added the change to package.json to point to my repo, but not tested
that part (I used npm link).